Move RowMap class and Associated Traits to own file, move DarwinCore and Symbiota Native Class to own file
