+++ app/Core/Download/SymbiotaNative.php
@@ -1,0 +1,132 @@
+<?php
+namespace App\Core\Download;
+
+class SymbiotaNative {
+    use RowMap;
+    use DeriveOccurrenceReference;
+    use DeriveCombineOccurrenceRecordID;
+
+    static $select = [''];
+
+    static $casts = [
+        'occid' => 'id',
+        'tidInterpreted' => 'taxonID',
+        'taxonRank' => 'verbatimTaxonRank',
+        'sourcePrimaryKey-dbpk' => 'dbpk',
+        'dateLastModified' => 'modified'
+    ];
+
+    static $ignores = [];
+
+    static $fields = [
+        'id' => '',
+        'institutionCode' => '',
+        'collectionCode' => '',
+        'ownerInstitutionCode' => '',
+        'basisOfRecord' => '',
+        'occurrenceID' => '',
+        'catalogNumber' => '',
+        'otherCatalogNumbers' => '',
+        'higherClassification' => '',
+        'kingdom' => '',
+        'phylum' => '',
+        'class' => '',
+        'order' => '',
+        'subgenus' => '',
+        'family' => '',
+        'scientificName' => '',
+        'taxonID' => '',
+        'scientificNameAuthorship' => '',
+        'genus' => '',
+        'specificEpithet' => '',
+        'verbatimTaxonRank' => '',
+        'infraspecificEpithet' => '',
+        'cultivarEpithet' => '',
+        'tradeName' => '',
+        'taxonRank' => '',
+        'identifiedBy' => '',
+        'dateIdentified' => '',
+        'identificationReferences' => '',
+        'identificationRemarks' => '',
+        'taxonRemarks' => '',
+        'identificationQualifier' => '',
+        'typeStatus' => '',
+        'recordedBy' => '',
+        'associatedCollectors' => '',
+        'recordNumber' => '',
+        'eventDate' => '',
+        'eventDate2' => '',
+        'year' => '',
+        'month' => '',
+        'day' => '',
+        'startDayOfYear' => '',
+        'endDayOfYear' => '',
+        'verbatimEventDate' => '',
+        'occurrenceRemarks' => '',
+        'habitat' => '',
+        'substrate' => '',
+        'verbatimAttributes' => '',
+        'behavior' => '',
+        'vitality' => '',
+        'fieldNumber' => '',
+        'eventID' => '',
+        'informationWithheld' => '',
+        'dataGeneralizations' => '',
+        'dynamicProperties' => '',
+        'associatedOccurrences' => '',
+        'associatedSequences' => '',
+        'associatedTaxa' => '',
+        'reproductiveCondition' => '',
+        'establishmentMeans' => '',
+        'cultivationStatus' => '',
+        'lifeStage' => '',
+        'sex' => '',
+        'individualCount' => '',
+        'samplingProtocol' => '',
+        'preparations' => '',
+        'locationID' => '',
+        'continent' => '',
+        'waterBody' => '',
+        'islandGroup' => '',
+        'island' => '',
+        'country' => '',
+        'countryCode' => '',
+        'stateProvince' => '',
+        'county' => '',
+        'municipality' => '',
+        'locality' => '',
+        'locationRemarks' => '',
+        'localitySecurity' => '',
+        'localitySecurityReason' => '',
+        'decimalLatitude' => '',
+        'decimalLongitude' => '',
+        'geodeticDatum' => '',
+        'coordinateUncertaintyInMeters' => '',
+        'verbatimCoordinates' => '',
+        'georeferencedBy' => '',
+        'georeferenceProtocol' => '',
+        'georeferenceSources' => '',
+        'georeferenceVerificationStatus' => '',
+        'georeferenceRemarks' => '',
+        'minimumElevationInMeters' => '',
+        'maximumElevationInMeters' => '',
+        'minimumDepthInMeters' => '',
+        'maximumDepthInMeters' => '',
+        'verbatimDepth' => '',
+        'verbatimElevation' => '',
+        'disposition' => '',
+        'language' => '',
+        'recordEnteredBy' => '',
+        'modified' => '',
+        'sourcePrimaryKey-dbpk' => '',
+        'collid' => '',
+        'recordID' => '',
+        'references' => '',
+    ];
+
+    static $derived = [
+        'references' => 'derive_references',
+        'recordID' => 'derive_combine_occurrence_record_id',
+        'occurrenceID' => 'derive_combine_occurrence_record_id'
+    ];
+}
